+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 Core
 ~~~~
 
+- Add support for service behavior costs on post-tax totals
 - Create separate refund lines for quantities and amounts
 - Fix handling of refunds for discounted lines